% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{number of vector of numbers}
}
\value{
Number of decimal places to which the number is rounded.
}
\description{
How many decimal places is a number rounded to?
}
\examples{
decimalplaces(1.2) #1
decimalplaces(702) #0
decimalplaces(3.14159) #5
}

