% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listObjects}
\alias{listObjects}
\title{list objects & sizes in R session}
\usage{
listObjects(..., n = 10)
}
\arguments{
\item{...}{generally unused (see source to \code{.ls.objects})}

\item{n}{max number of objects to list (default 10)}
}
\value{
data frame of object information
}
\description{
list objects & sizes in R session
}
\examples{
\dontrun{
listObjects()
}
}

