% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makepretty}
\alias{makepretty}
\title{Make a string pretty!}
\usage{
makepretty(x)
}
\arguments{
\item{x}{string (i.e., character vector of length 1)}
}
\value{
string that is the same as \code{x} but with all newline (\code{'\n'}) and 4-space blocks
  (\code{'    '}) replaced with the empty string.
}
\description{
strips newlines and blocks of 4 spaces from strings. I use this when I want to have
  newlines and 4-space tabs in strings *in my script*, but would also like to print them to
  the console in a sensible way.
}
\examples{
\dontrun{
msg = 'Hello, I will type a message to you
    like this!'
message(msg)
message(makepretty(msg))
}
}

