% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTransparent}
\alias{makeTransparent}
\title{get transparent equivalents of R colors}
\usage{
makeTransparent(..., alpha = 0.5)
}
\arguments{
\item{...}{any number of colors, in any of R's acceptable forms (hex, character, number, etc)}

\item{alpha}{how transparent do you want it? must be between 0 (totally transparent) and 1
(totally opaque)}
}
\value{
vector of transparentized colors, same length as \code{...}
}
\description{
get transparent equivalents of R colors
}

