% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_cols.R
\name{nice_cols}
\alias{nice_cols}
\title{Nice colum names}
\usage{
nice_cols(string)
}
\arguments{
\item{columns}{removes non character and numeric values from columns}
}
\description{
Nice colum names
}
\examples{
colnames(iris) <- nice_cols(string = iris)
}
