% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_cols_df.R
\name{nice_cols_df}
\alias{nice_cols_df}
\title{Nice col names over data frame (pipeable)}
\usage{
nice_cols_df(.data)
}
\arguments{
\item{dataframe}{to rename columns in dataframe as part of a piped opetation}
}
\value{
dataframe
}
\description{
Nice col names over data frame (pipeable)
}
\examples{
nice_cols_df(iris)
}
