% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ss}
\alias{ss}
\title{ss}
\usage{
ss(x, pattern, slot = 1, ...)
}
\arguments{
\item{x}{character vector}

\item{pattern}{slot separator}

\item{slot}{slot index to get (default 1)}

\item{...}{extra arguments for \code{strsplit}}
}
\value{
vector, same length as \code{x}, with elements consisting of only
the \code{slot}th slot of the corresponding elements in \code{x}.
}
\description{
extract a specific slot from each element of a character vector
}
\examples{
fullnames = c('Hilary Clinton', 'Florence Nightingale', 'Rosalind Franklin')
  firstnames = ss(fullnames, ' ', slot=1)
  lastnames = ss(fullnames, ' ', slot=2)
}

