% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strip_quotes}
\alias{strip_quotes}
\title{strip_quotes}
\usage{
strip_quotes(x)
}
\arguments{
\item{x}{character vector}
}
\value{
\code{x}, but with first and last characters stripped from all
elements
}
\description{
strip off the first and last characters of each element of a
character vector
}
\details{
I've always found this useful with character vectors containing
quoted elements (e.g. if quoting was inconsistent in the file you read the
vector from).
}
\examples{
gene_names = c('"OCT4"', '"BRCA1"', '"ZZZ3"')
genes_noquote = strip_quotes(gene_names)
}

