% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bhist}
\alias{bhist}
\title{overlay histograms!}
\usage{
bhist(z, add = FALSE, breaks = 20, col = "black", fill = "orange",
  alpha = 0.5, xlim = c(-0.01, 1.01), ylim = NULL, freq = TRUE, ...)
}
\arguments{
\item{z}{numeric, vector of values for desired histogram}

\item{add}{if TRUE, overlay histogram on current graphics device}

\item{breaks}{see \code{breaks} in \code{hist()}}

\item{col}{line color for histogram's bins}

\item{fill}{fill color for histogram}

\item{freq}{whether to make a frequency or density histogram. (TRUE for frequency)}
}
\description{
base graphics implementation of overlaid histograms
}

